excel

+++ SpringBoot/ticketlog-api/src/main/java/com/ticketlogapi/controller/CidadeController.java
@@ -1,7 +1,10 @@
 package com.ticketlogapi.controller;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
+
+import javax.persistence.EntityNotFoundException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.ticketlogapi.entities.Cidade;
+import com.ticketlogapi.entities.Estado;
 import com.ticketlogapi.services.CidadeService;
 import com.ticketlogapi.services.CustosService;
 import com.ticketlogapi.services.EstadoService;
@@ -37,15 +41,17 @@ public class CidadeController {
 		return ResponseEntity.ok().body(service.findAll());
 	}
 
-
 	@GetMapping(path = "/{name}")
 	public ResponseEntity<List<Cidade>> findByName(@PathVariable String name) {
 		return ResponseEntity.ok().body(service.findByName(name));
 	}
 
-	@PostMapping
-	public ResponseEntity<Cidade> addCidade(@RequestBody Cidade cidade) {
+	@PostMapping(path = "/{estado}")
+	public ResponseEntity<Cidade> addCidade(@PathVariable String estado, @RequestBody Cidade cidade) {
+		Estado estados = service.cidadePorEstado(estado);
+		cidade.setEstado(estados);
 		service.addCidade(cidade);
+		
 		custosService.CalculaCusto(cidade.getId());
 		estadoService.sumPopulacao(cidade.getEstadoId());
 		custosService.CustoEstado(cidade.getId());
@@ -54,8 +60,13 @@ public class CidadeController {
 
 	@DeleteMapping(path = "/{id}")
 	public ResponseEntity<Cidade> deleteById(@PathVariable int id) {
-		service.deleteCidade(id);
-		return new ResponseEntity<>(HttpStatus.OK);
+		if (service.cidadeRS(id)) {
+			return new ResponseEntity<>(HttpStatus.METHOD_NOT_ALLOWED);
+		} else {
+			service.deleteCidade(id);
+			return new ResponseEntity<>(HttpStatus.OK);
+		}
+
 	}
 
 }
